+++ backend/backend/urls.py
@@ -17,7 +17,7 @@ from django.contrib import admin
 from django.urls import include, path, re_path
 from rest_framework import routers
 
-from clover_backend.games import views
+from backend.games import views
 
 admin.autodiscover()
 from rest_framework.routers import SimpleRouter
