--- backend/clover_backend/clover/admin.py
@@ -1,3 +1,0 @@
-from django.contrib import admin
-
-# Register your models here.
